Added logDebug(OMElement om,Log log...) methods to CommonUtils.
The logDebug(...) is a convenience method that serializes the OMElement to a special output stream, LogOutputStream,
which logs the message.  The implementation ensures that the entire message is not stored in-core; thus it reduces
the chances of an OutOfMemory condition.

A side effect of logDebug is that it returns the length of the message.