Simplified DOMStAXWrapper#getAttributeLocalName by removing the if statement in the following piece of code:
if (attrib.getNamespace() != null) {
    returnString = attrib.getLocalName();
} else {
    returnString = ((Attr) attrib).getNodeName();
}

The methods in AttrImpl that are used by this code are defined as follows: 

public OMNamespace getNamespace() {
    return this.namespace;
}

public String getLocalName() {
    return (this.namespace == null) ? this.attrName : DOMUtil
            .getLocalName(this.attrName);
}

public String getNodeName() {
    return (this.namespace != null
            && !"".equals(this.namespace.getPrefix()) &&
            !(OMConstants.XMLNS_NS_PREFIX.equals(this.attrName)))
            ? this.namespace.getPrefix() + ":" + this.attrName
            : this.attrName;
}

It can be seen that if getNamespace returns null, then both getLocalName and getNodeName return the same value (attrName). Therefore the code shown above can be simply rewritten as:

returnString = attrib.getLocalName();

This is important because this code is the only really DOOM specific piece in DOMStAXWrapper.