Merged the code in OMStAXWrapper and DOMStAXWrapper into a single base class.
Rationale: There is nothing implementation (LLOM/DOOM) specific in this code and it doesn't make sense to maintain two separate copies of it (with over 1500 lines of code).

Note that the only remaining differences between these two classes were:
* Code handling OMSourcedElements (which don't exist in DOOM).
* Code related to WSCOMMONS-220 (StAX getProperty(String) support and keep track of parser closure state).
* Code related to WSCOMMONS-280 (Control string interning of OMStAXWrapper.getNamespaceURI()).
* Code related to WSCOMMONS-344 (OMStaXWrapper Does Not Support Optimized OMText as an XOP:Include).
* Some additional null checks and logging. 
All this should also be available to DOOM.

This change has been tested successfully with Axis2 trunk and Rampart trunk.