Override getOMFactory in OMDocumentImpl.  This allows OMDocumentImpl constructors that don't take an OMFactory parameter to return a valid OMFactory.
Revision 786706 refactored getOMFactory into superclass OMSerializableImpl.  The other subclass, OMNodeImpl, always takes an OMFactory as a parameter to its constructors, so it doesn't have the same problem.  OMDocumentImpl has several constructors such as OMDocumentImpl(OMXMLParserWrapper) that do not take have a factory parameter so the value can not be set in the superclass during construction.  For those code paths, the factory needs to be gotten at a later time.