Moved some of the iterator implementations from axiom-api to axiom-common-impl, while keeping deprecated copies in axiom-api.
Rationale: Normally these classes are not considered part of the public API (application code is supposed to use the corresponding methods in OMContainer/OMElement). Having them in axiom-common-impl means that we can freely refactor them without the risk of breaking existing code. The deprecated classes will be removed in Axiom 1.3.