Dead code elimination:
If OMNavigator#isNavigable() and OMNavigator#isCompleted() both return false, then this means that the next node has not been built. This implies that there must be a builder and that the builder is not complete, i.e. builder != null and StAXOMBuilder#isCompleted() returns false.

Note that this code was added in AXIOM-116, but no unit test was ever produced to show the purpose of this code.