AXIOM-502: make AXIOM usable in GraalVM Native Image
When building an application using the GraalVM Native Image compiler it
fails when AXIOM is used in the application. The GraalVM compiler fails
with the exception java.lang.InternalError: Enclosing method not found.

This seems to be caused by missing type information for the anonymous
AbstractNodeIterator implementation in CoreParentNodeSupport.

By replacing the anonymous inner class with a concrete implementation
AXIOM can be used with the GraalVM Native Image compiler.