AXIOM-506: Remove the axiom-all module
The Uber-JAR doesn't add much value because it is not deployed to Maven
(for good reasons), but nowadays everyone should be using Maven
repositories to fetch dependencies. On the other hand, the fact that
some of our APIs now have two versions (one for javax and one for
Jakarta) makes maintaining an Uber-JAR more complicated.